+++ dart/geobase/lib/src/vector_data/model/feature/feature.dart
@@ -254,9 +254,7 @@ class Feature<T extends Geometry> extends FeatureObject {
       geometry: geom,
       properties: properties,
       custom: custom,
-      bounds: recalculate || bounds == null
-          ? _buildBoundsFrom(geom)
-          : bounds,
+      bounds: recalculate || bounds == null ? _buildBoundsFrom(geom) : bounds,
     );
   }