+++ dart/geobase/test/vector/wkb_test.dart
@@ -34,7 +34,7 @@ void main() {
             // three different ways to write POINT(2.0 4.0)
             (writer) => writer.point([2.0, 4.0].xy),
             (writer) =>
-                writer.point(const Geographic(lon: 2.0, lat: 4.0).coords()),
+                writer.point(const Geographic(lon: 2.0, lat: 4.0).coords),
             (writer) => writer.point([2.0, 4.0])
           ],
         );
@@ -50,7 +50,7 @@ void main() {
             // four different ways to write empty point
             (writer) => writer.point([double.nan, double.nan].xy),
             (writer) => writer.point(
-                  const Geographic(lon: double.nan, lat: double.nan).coords(),
+                  const Geographic(lon: double.nan, lat: double.nan).coords,
                 ),
             (writer) => writer.point([double.nan, double.nan]),
             (writer) => writer.emptyGeometry(Geom.point),
@@ -66,7 +66,7 @@ void main() {
             // three different ways to write POINT(2.1 -3.4)
             (writer) => writer.point([2.1, -3.4].xy),
             (writer) =>
-                writer.point(const Geographic(lon: 2.1, lat: -3.4).coords()),
+                writer.point(const Geographic(lon: 2.1, lat: -3.4).coords),
             (writer) => writer.point([2.1, -3.4])
           ],
         );