+++ dart/geobase/CHANGELOG.md
@@ -4,6 +4,7 @@ NOTE: Version 0.6.0 currently [under development](https://github.com/navibyte/ge
 
 🧩 Features:
 * [Add copyWith (and map) methods to coordinate, meta, feature and geometry classes #189](https://github.com/navibyte/geospatial/issues/189)
+* [Allow a position stored as "Position" in Point, and a bounding box as "Box" in Geometry classes #192](https://github.com/navibyte/geospatial/issues/192)
 
 ⚠️ Breaking changes:
 * [Simplify Feature and FeatureCollection types #195](https://github.com/navibyte/geospatial/issues/195)